requirements.txt
diff-match-patch==20121119
Django==2.0.8
django-bootstrap4==0.0.6
django-import-export==1.0.1
et-xmlfile==1.0.1
Faker==0.9.2
gunicorn==19.9.0
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.5
psycopg2-binary==2.7.5
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
six==1.11.0
tablib==0.12.1
text-unidecode==1.2
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
